sistemaEmprestimo/requirements.txt
asgiref==3.2.10
asn1crypto==1.3.0
bcrypt==3.1.7
cffi
cryptography
dj-database-url==0.5.0
Django==3.1.2
django-bootstrap3==12.0.3
django-mail-templated==2.6.5
django-widget-tweaks==1.4.8
gunicorn==20.0.4
idna==2.9
invoke==1.4.1
paramiko==2.7.1
pyasn1==0.4.8
pycparser==2.20
PyNaCl
python-decouple==3.8
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
